+++ include/common/constants.h
@@ -12,6 +12,9 @@ static constexpr GInt BASE2 = 2;
 static constexpr GDouble HALF = 0.5;
 static constexpr GInt INVALID_CELLID = -1;
 
+/// Generate invalid list for init of arrays
+/// \tparam LENGTH Length of the invalid list
+/// \return Invalid list of LENGTH
 template <GInt LENGTH>
 static constexpr auto INVALID_LIST() -> std::array<GInt, LENGTH> {
   std::array<GInt, LENGTH> invalid{};
@@ -28,7 +31,7 @@ namespace timeconst {
 static constexpr GInt MINUTE = 60;
 static constexpr GDouble DMINUTE = 60;
 static constexpr GInt HOUR = 3600;
-static constexpr GDouble DHOUR   = 3600;
+static constexpr GDouble DHOUR = 3600;
 static constexpr GInt    DAY     = HOUR * 24;
 static constexpr GDouble DDAY    = DHOUR * 24;
 static constexpr GInt    WEEK    = DAY * 7;