+++ include/common/math/hilbert.h
@@ -13,15 +13,16 @@ namespace hilbert {
 /// hilbertLevel The number of iterations of the Hilbert curve. \return The
 /// Hilbert index. Which should be unique for cells in a valid Cartesian grid.
 template <GInt NDIM>
-inline auto index(const VectorD<NDIM>& x, const GInt hilbertLevel) -> GInt {
+inline auto index(const VectorD<NDIM> &x, const GInt hilbertLevel) -> GInt {
   // todo: make this assert work
-  //    ASSERT(static_cast<GBool>(x.array() >=0) && static_cast<GBool>(x.array() <=1), "Invalid Coordinates");
+  //    ASSERT(static_cast<GBool>(x.array() >=0) && static_cast<GBool>(x.array()
+  //    <=1), "Invalid Coordinates");
   VectorD<NDIM> position = x;
-  GInt          index    = 0;
+  GInt index = 0;
 
-  for(GInt level = 0; level < hilbertLevel; ++level) {
+  for (GInt level = 0; level < hilbertLevel; ++level) {
     std::bitset<NDIM> quadrant;
-    for(GInt dir = 0; dir < NDIM; ++dir) {
+    for (GInt dir = 0; dir < NDIM; ++dir) {
       quadrant[dir] = position[dir] >= 0.5;
     }
     const GInt hilbertLUTId = quadrant.to_ulong();