+++ include/common/timer.h
@@ -4,7 +4,6 @@
 #define SFCMM_TIMER_H
 
 #include "sfcmm_types.h"
-#include "term.h"
 #include <chrono>
 #include <cmath>
 #include <common/math/mathfunctions.h>
@@ -15,6 +14,9 @@
 #include <unistd.h>
 #include <utility>
 #include <vector>
+
+//todo: documentation
+//todo: tests
 
 using chronoTimePoint = std::chrono::time_point<std::chrono::system_clock, std::chrono::duration<double>>;
 
@@ -236,9 +238,8 @@ inline void TimerManager::recordTimerStart(const GInt timerId, const GString& po
 
 inline void TimerManager::startTimer(const GInt timerId, const GString& pos) {
   ASSERT(m_timers[timerId].status() != Timer::Running,
-         "The timer " + m_timers[timerId].name() +
-             " with id: " + std::to_string(timerId) +
-             " can't be started because it is already running! " + pos);
+         "The timer " + m_timers[timerId].name + " with id: " + std::to_string(timerId)
+             + " can't be started because it is already running! " + pos);
   std::ignore = pos;
 
   chronoTimePoint t = time();