+++ include/sfcmm_common.h
@@ -15,6 +15,8 @@ static constexpr int MAX_DIM = 4;
 #include "common/timer.h"
 
 #include "common/util/backtrace.h"
+#include "common/util/base64.h"
+#include "common/util/binary.h"
 #include "common/util/string_helper.h"
 #include "common/util/sys.h"